Signed Uri response
When developing locally I use Minio as my s3 service. I have this running on port 9000.

With the current setup I'd have to override the whole `store` method just to change the url in the response.

To the get URL from the signed request you can cast the Uri to a string to get what you need instead of building the URL manually.